added groupController test mockBean InstructorService
Took 54 minutes

+++ src/main/java/io/codelex/studentsystem/StudentController.java
@@ -3,6 +3,8 @@ package io.codelex.studentsystem;
 import io.codelex.studentsystem.api.Student;
 import io.codelex.studentsystem.api.requests.AddStudent;
 import io.codelex.studentsystem.repository.service.StudentService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -22,13 +24,22 @@ public class StudentController {
     }
 
     @GetMapping("/internal-api/students/{id}")
-    public Student findStudent(@PathVariable("id") Long id) {
-        return service.findStudentById(id);
+    public ResponseEntity<Student> findStudent(@PathVariable("id") Long id) {
+        if (service.findStudentById(id) != null) {
+            return new ResponseEntity<>(service.findStudentById(id), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
     @DeleteMapping("/internal-api/students/{studentsId}")
-    public void deleteById(@PathVariable long studentsId) {
-        service.deleteById(studentsId);
+    public ResponseEntity<String> deleteById(@PathVariable long studentsId) {
+        if (service.findStudentById(studentsId) != null) {
+            service.deleteById(studentsId);
+            return new ResponseEntity<>("Student deleted", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(studentsId + " id already does not exist!", HttpStatus.BAD_REQUEST);
+        }
     }
 
 }
